Update pinpoint notification test
